pyproject.toml
[tool.poetry]
name = "shop-site"
version = "0.1.0"
description = ""
authors = ["Romankrasawa <romankrasawa8567@gmail.com>"]
readme = "README.md"
packages = [{include = "shop_site"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.4"
psycopg2 = "^2.9.5"
python-slugify = "^7.0.0"
pillow = "^9.4.0"
gunicorn = "^20.1.0"
python-dotenv = "^0.21.0"
django-redis = "^5.2.0"
django-allauth = "^0.52.0"
requests = "^2.28.2"
celery = "^5.2.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

